options-bot/requirements.txt
# Core framework
lumibot>=4.4.50
alpaca-py>=0.43.0

# ML - Phase 1
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
ta>=0.11.0

# Backend
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0

# ML - Phase 4 (installed now to avoid dependency surprises later)
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
tensorboard>=2.14.0
